+++ catalogs/mingcute.ts
@@ -3,31 +3,15 @@ import { fetchIcon } from "./utils.ts";
 
 const assets = "https://cdn.jsdelivr.net/gh/Richard9394/MingCute@2.95";
 
-export type Type = "fill" | "line";
-
-export interface Options {
-  defaultType: Type;
-}
-
-export const defaults: Options = {
-  defaultType: "fill",
-};
-
-export default function (options: Options = defaults) {
-  return new MingCute(options);
+export default function () {
+  return new MingCute();
 }
 
 export class MingCute implements Catalog<void> {
   name = "mingcute";
-  defaultType: Type;
 
-  constructor(options: Options = defaults) {
-    this.defaultType = options.defaultType;
-  }
-
-  get(name: string, type?: Type): Promise<string> {
-    const iconType = type || this.defaultType;
-    const url = `${assets}/svg/${name}_${iconType}.svg`;
+  get(name: string): Promise<string> {
+    const url = `${assets}/svg/${name}.svg`;
     return fetchIcon(url);
   }